host: Make log aggregation highly available and recoverable
This rewrites the logmux package and replaces the logbuf package.

Logs are now saved to local files in a slightly modified syslog format.

logaggregator instances are found via a discoverd watch, and when they
come up, the most recent host cursor is determined, and the logs for each
app since that cursor are replayed.

With this change if a logaggregator instance comes up with no data in
its buffers, all hosts that are up will replay the relevant logs to it.

Signed-off-by: Jonathan Rudenberg <jonathan@titanous.com>